fix(blog): move Popular Topics nebula into the sidebar
The nebula was a full-width band above the post grid. Move it into the
blog sidebar, into the slot the old "Popular tag" widget occupied (after
Recent Posts), so it sits where popular tags always lived rather than on
top of everything. Orb size is now derived from the measured column width
so the nebula fits the narrow sidebar. Relocated the component to
sidebar/widgets and removed it from BlogsMain.
