+++ src/components/shared/sidebar/BlogSidebar.js
@@ -4,6 +4,7 @@ import Link from "next/link";
 import { useRouter, useSearchParams } from "next/navigation";
 import { useRef, useState } from "react";
 import BlogCategoriesWidget from "./widgets/BlogCategoriesWidget";
+import PopularTagsNebula from "./widgets/PopularTagsNebula";
 import RecentBlogWidget from "./widgets/RecentBlogWidget";
 import makePath from "@/libs/makePath";
 import makeText from "@/libs/makeText";
@@ -76,6 +77,8 @@ const BlogSidebar = () => {
         <BlogCategoriesWidget />
         {/* <!-- recent blogs--> */}
         <RecentBlogWidget />
+        {/* <!-- popular topics --> */}
+        <PopularTagsNebula />
       </div>
     </div>
   );
